requirements.txt
appdirs==1.4.4
asgiref==3.2.8
astroid==2.4.2
boto3==1.16.10
botocore==1.19.10
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-cleanup==5.1.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-phone-field==1.8.1
django-storages==1.10.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
isort==5.4.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.0
psycopg2-binary==2.8.5
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.11
virtualenv==20.0.23
whitenoise==5.2.0
wrapt==1.12.1
